# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo List Api",
        "description": "Demo Todo List Api",
        "version": "1.0"
    },
    "paths": {
...
        "/api/delete-taskbyid": {
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete task by id",
                "operationId": "4ec7eec25a279b1ed8b279e87251308f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "delete successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/addtask": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Add task",
                "operationId": "b23772a523efa7a4a114414b0a8ad81d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "completed": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "important": {
                                        "type": "number",
                                        "example": "0"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "taskname"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "day src len github"
                                    },
                                    "date": {
                                        "type": "string",
                                        "example": "2023-06-09"
                                    },
                                    "directory": {
                                        "type": "string",
                                        "example": "main"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Model created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}